requirements.txt
flask-login
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3

SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1

psycopg2-binary==2.9.11

redis==5.0.8
rq==2.6.1

cryptography==43.0.3

openai==2.15.0
cerebras-cloud-sdk==1.64.1

pytesseract==0.3.13
Pillow==11.1.0

firebase-admin